full-stack-forum-user-service/requirements.txt
flask
flask-restful
flask-sqlalchemy
flask-jwt-extended
flask-cors
flask-bcrypt
python-dotenv
dotenv
Werkzeug
redis
pika
validate-email-address
pymysql